# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TarotGen.io API",
        "description": "The backend for **TarotGen.io** — a PHP 8.5 / Slim 4 hybrid-REST API.\n\n- Resources are plural nouns (`/decks`, `/readings`); items are `/{id}`.\n- `GET` reads, `POST` creates (→`201`) or runs a command, `PUT` replaces,\n  `PATCH` applies a partial update / flag toggle, `DELETE` removes (→`204`).\n- State transitions that aren't a plain field-set are `POST …/{id}/{verb}`\n  (e.g. `…/finalize`, `…/approve`, `…/draw`).\n- Errors are `{ \"error\": \"…\" }`; multi-message validation is `{ \"errors\": […] }`.\n- Auth is a session cookie (`PHPSESSID`) set by the `/auth/*` endpoints —\n  there are no API keys or bearer tokens for first-party calls.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Same-origin API base path"
        }
    ],
    "paths": {
...
        "/readings/{reading_id}": {
            "get": {
                "tags": [
                    "Readings"
                ],
                "summary": "Fetch a reading (password-protected readings return a locked stub to non-owners)",
                "operationId": "e7982bebe092fc207c5d04e2c203ecba",
                "parameters": [
                    {
                        "name": "reading_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The reading, or a { locked, reading_name } stub",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reading"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "InvalidReadingID"
                    }
                }
            }
        },
        "/readings/{reading_id}/unlock": {
            "post": {
                "tags": [
                    "Readings"
                ],
                "summary": "Unlock a password-protected reading for this session",
                "operationId": "510c945a2fb0ffe6c221f7483e01f021",
                "parameters": [
                    {
                        "name": "reading_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The unlocked reading",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reading"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Incorrect password"
                    },
                    "404": {
                        "description": "InvalidReadingID"
                    }
                }
            }
        },
        "/readings/generate": {
            "post": {
                "tags": [
                    "Readings"
                ],
                "summary": "Generate a random reading from a draw spec",
                "operationId": "d8750b5a23fd03c315941e4facfd28ee",
                "requestBody": {
                    "description": "Draw spec: deck, spread or card count, and options",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created reading",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reading"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid draw spec"
                    }
                }
            }
        },
        "/readings": {
            "post": {
                "tags": [
                    "Readings"
                ],
                "summary": "Create a custom reading from explicitly chosen cards/positions",
                "operationId": "3e30d906e6fde22906f101631e2f600d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created reading",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reading"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/readings/{reading_id}/placement": {
            "put": {
                "tags": [
                    "Readings"
                ],
                "summary": "Save spread/placement positions onto an existing reading",
                "description": "Update a reading's spread/placement information after a \"Free Draw With Placement\"\nhas been generated. The caller provides position data which is merged into the\nexisting reading_info JSON as a spread snapshot.",
                "operationId": "36bf5ca46d89e9e49656c8a282962abc",
                "parameters": [
                    {
                        "name": "reading_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "positions"
                                ],
                                "properties": {
                                    "positions": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "spread_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "reading_id": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Position count mismatch or invalid data"
                    },
                    "403": {
                        "description": "Not the owner"
                    },
                    "404": {
                        "description": "Reading not found"
                    },
                    "409": {
                        "description": "Reading is final"
                    }
                }
            }
        },
        "/readings/{reading_id}/draw": {
            "post": {
                "tags": [
                    "Readings"
                ],
                "summary": "Draw additional cards into a non-final, non-custom reading you own",
                "description": "Draw additional cards into an existing generated reading owned by the\ncaller. The new cards are appended to the draw; for a spread reading the\nclient then opens the placement editor to position them.",
                "operationId": "add5b38b9a8b679c94913d175f42e088",
                "parameters": [
                    {
                        "name": "reading_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated reading",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reading"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Custom readings cannot draw"
                    },
                    "403": {
                        "description": "Not the owner"
                    },
                    "404": {
                        "description": "Reading not found"
                    },
                    "409": {
                        "description": "Reading is final"
                    }
                }
            }
        },
        "/readings/{reading_id}/finalize": {
            "post": {
                "tags": [
                    "Readings"
                ],
                "summary": "Permanently lock a reading against further draws (owner-only, one-way)",
                "description": "Mark a reading as final, permanently locking it against further draws.\nOne-way: there is no endpoint to undo it. Owner-only.",
                "operationId": "590c79f6c9fbc27357aae420968f729c",
                "parameters": [
                    {
                        "name": "reading_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The finalized reading",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reading"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not the owner"
                    },
                    "404": {
                        "description": "Reading not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reading": {
                "description": "A saved tarot reading.",
                "properties": {
                    "reading_id": {
                        "type": "string"
                    },
                    "reading_info": {
                        "description": "Serialized to a nested JSON object in the response (see jsonSerialize).",
                        "type": "object"
                    },
                    "reading_time": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "hide_user": {
                        "type": "boolean"
                    },
                    "reading_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "reading_notes": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "is_final": {
                        "description": "One-way lock: when true the owner can no longer draw additional cards.",
                        "type": "boolean"
                    },
                    "password_protected": {
                        "description": "Derived flag — true when a view password is set. The hash itself is never exposed.",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "description": "Session cookie issued on sign-in. Sent automatically by the browser.",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}